Use same component for all edits of mission schedule
